Fix review issues: dedup enum, fix duplicate call, clean up CI
- Remove duplicate resolve_host_var_list() call in esql.c
- Create esql-common.h to share cb_sql_command enum definition
  with #ifndef guard, eliminating duplication between tree.h
  and esql-parser.y
- Clean up empty lines in test-esql.yml, consolidate test steps
- Add sample.at to esql_basic_DEPENDENCIES in Makefile.am
