+++ .github/workflows/test-esql.yml
@@ -85,40 +85,14 @@ jobs:
           mv ${TEMP_FILE} $1
           DBEOF
 
-      - name: Run ESQL basic tests
-        id: esql-basic
-
-
+      - name: Run ESQL tests
         working-directory: tests/
-        run: ./esql-basic
-
-      - name: Run ESQL SQLCA tests
-        id: esql-sqlca
-
-
-        working-directory: tests/
-        run: ./esql-sqlca
-
-      - name: Run ESQL cobol data tests
-        id: esql-cobol-data
-
-
-        working-directory: tests/
-        run: ./esql-cobol-data
-
-      - name: Run ESQL misc tests
-        id: esql-misc
-
-
-        working-directory: tests/
-        run: ./esql-misc
-
-      - name: Run ESQL sql data tests
-        id: esql-sql-data
-
-
-        working-directory: tests/
-        run: ./esql-sql-data
+        run: |
+          ./esql-basic
+          ./esql-sqlca
+          ./esql-cobol-data
+          ./esql-misc
+          ./esql-sql-data
 
       - name: Upload test logs if tests fail
         if: failure()
