+++ cobj/esql-parser.y
@@ -8,54 +8,7 @@
 %{
 #include <stdlib.h>
 #include <string.h>
-
-/* Forward declarations from tree.h (avoid including tree.h directly
-   because it defines YYSTYPE as cb_tree, conflicting with ESQL_STYPE) */
-struct cb_tree_common;
-typedef struct cb_tree_common *cb_tree;
-extern cb_tree cb_error_node;
-extern cb_tree cb_build_reference(const char *name);
-
-struct cb_sql_host_var;
-extern struct cb_sql_host_var *cb_build_sql_host_var(char *name, cb_tree ref);
-extern struct cb_sql_host_var *
-cb_sql_host_var_list_add(struct cb_sql_host_var *list,
-                         struct cb_sql_host_var *item);
-
-enum cb_sql_command {
-  CB_SQL_CONNECT,
-  CB_SQL_CONNECT_INFORMAL,
-  CB_SQL_CONNECT_SHORT,
-  CB_SQL_DISCONNECT,
-  CB_SQL_EXEC,
-  CB_SQL_EXEC_PARAMS,
-  CB_SQL_SELECT_INTO_ONE,
-  CB_SQL_SELECT_INTO_OCCURS,
-  CB_SQL_DECLARE_CURSOR,
-  CB_SQL_DECLARE_CURSOR_PARAMS,
-  CB_SQL_OPEN_CURSOR,
-  CB_SQL_OPEN_CURSOR_PARAMS,
-  CB_SQL_CLOSE_CURSOR,
-  CB_SQL_FETCH_ONE,
-  CB_SQL_FETCH_OCCURS,
-  CB_SQL_PREPARE,
-  CB_SQL_EXECUTE_PREPARED,
-  CB_SQL_COMMIT,
-  CB_SQL_ROLLBACK,
-};
-
-extern void cb_error(const char *msg, ...);
-
-/* Forward declarations */
-extern const char *esql_get_sqlbody(void);
-extern cb_tree esql_build_and_resolve(enum cb_sql_command command,
-                                      char *sql_text, char *cursor_name,
-                                      char *prepare_name, char *db_name,
-                                      struct cb_sql_host_var *host_list,
-                                      int host_count,
-                                      struct cb_sql_host_var *res_host_list,
-                                      int res_host_count,
-                                      int conn_use_other_db);
+#include "esql-common.h"
 
 int esql_lex(void);
 void esql_error(const char *msg);
